Rename "migrate" command to "sync"
