+++ lib/sync/index.js
