+++ lib/sync/util/apply.js
