+++ lib/sync/util/functions-migration.sql
