+++ lib/sync/util/resource-dependency-graph.js
